refactor: Move `TemplateEngineManager` under 📁`Engines/`
