+++ src/EleventyExtensionMap.js
@@ -1,6 +1,6 @@
 import { TemplatePath } from "@11ty/eleventy-utils";
 
-import TemplateEngineManager from "./TemplateEngineManager.js";
+import TemplateEngineManager from "./Engines/TemplateEngineManager.js";
 import EleventyBaseError from "./Errors/EleventyBaseError.js";
 
 class EleventyExtensionMapConfigError extends EleventyBaseError {}
